Added a script for enemy health and added damage to bullet collision with one zombie...for now.
Added a health bar to the zombie for test purposes.

There is an error in the code when it comees to the zombies object, for
the bullet damage to work I had to comment out line 56 in the
BulletScript.

This is the error that shows:
NullReferenceException: Object reference not set to an instance of an object
manZombieHealth.Start () (at Assets/scripts/enemy/manZombieHealth.cs:17)